% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{add_cluster_infos}
\alias{add_cluster_infos}
\title{Adds a 'cluster' id to a mutation data frame}
\usage{
add_cluster_infos(d, ccf_cutoff = 0.25, small_frac = 0.01)
}
\arguments{
\item{d}{A mutation data frame.}

\item{ccf_cutoff}{Cutoff of the CCF to consider sample mutated [default:0.25].}

\item{small_frac}{Relative size of the smallest cluster to label [default:0.01].}
}
\value{
A mutation data frame with a cluster id column 'clust_id' added.
}
\description{
Adds a 'cluster' id to a mutation data frame
}
