% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_annotation.R
\name{annotation_from_barcode}
\alias{annotation_from_barcode}
\title{Get sample annotations from sample barcodes.}
\usage{
annotation_from_barcode(barcodes, extract = FALSE)
}
\arguments{
\item{barcodes}{A character (or factor) vector of sample barcodes.}

\item{extract}{Logical flag if barcodes should be extracted from 'barcodes' (Default: False).}
}
\value{
A data frame with sample annotations, rownames are set if there are no duplicated barcodes.
}
\description{
Get sample annotations from sample barcodes.
}
\examples{
annotation_from_barcode("EPICC_C516_Z1_B1_D1") # a EPICC sample barcode
}
