% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicc_project.R
\name{annotation_from_barcode_epicc}
\alias{annotation_from_barcode_epicc}
\title{EPICC Project: Get sample annotations from sqmple barcodes.}
\usage{
annotation_from_barcode_epicc(barcodes, extract = FALSE)
}
\arguments{
\item{barcodes}{A character vector of EPICC sample barcodes.}

\item{extract}{Logical flag if barcodes should be extracted from 'barcodes' (Default: False).}
}
\value{
A data frame with sample annotations, rownames are set if there are no duplicated barcodes.
}
\description{
EPICC Project: Get sample annotations from sqmple barcodes.
}
\examples{
annotation_from_barcode_epicc("EPICC_C516_Z1_B1_D1")
}
