% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_project.R
\name{annotation_from_barcode_tcga}
\alias{annotation_from_barcode_tcga}
\title{TCGA project: Create a annotation for a vector of TCGA barcodes.}
\usage{
annotation_from_barcode_tcga(barcodes, extract = FALSE)
}
\arguments{
\item{barcodes}{A character vector of TCGA sample barcodes.}

\item{extract}{Logical flag if barcodes should be extracted from 'barcodes' (Default: FALSE).}
}
\value{
A data frame with sample annotations.
}
\description{
TCGA project: Create a annotation for a vector of TCGA barcodes.
}
\examples{
annotation_from_barcode_tcga("TCGA-02-0001-01C-01D")
}
