% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{baseRevComplement}
\alias{baseRevComplement}
\title{Complemental strand of a DNA sequence with preserved direction.}
\usage{
baseRevComplement(x)
}
\arguments{
\item{x}{DNA base sequence.}
}
\value{
Complemental strand with identical orientation (3'->5' and 3'->5').
}
\description{
Complemental strand of a DNA sequence with preserved direction.
}
\examples{
baseRevComplement("ACCCG")
}
\keyword{internal}
