% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{calc_pga_stats}
\alias{calc_pga_stats}
\title{Calculate PGA stats}
\usage{
calc_pga_stats(cna_data)
}
\arguments{
\item{cna_data}{Copy-number summary stats, must contain the following columns: sample_barcode, CN (the copy-numner), start (start of the CN interval), end (end of the CN interval), analyte_name (the type of the sequencing the data came from),  tissue_type (the type of the analysed tissue, e.g., cancer and adenoma), msi_status (MSI status or any additional tumour property).}
}
\value{
A list that contains PGA summary stats. Can be passed to plot_pga_stats to create a plot of these.
}
\description{
Calculate PGA stats
}
