% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{drop_multiallelic}
\alias{drop_multiallelic}
\title{Function to drop multiallelic sites from VCFs}
\usage{
drop_multiallelic(d)
}
\arguments{
\item{d}{Object of class 'CollapsedVCF'}
}
\value{
Object of class 'CollapsedVCF' with multiallelic sites split
}
\description{
Function to drop multiallelic sites from VCFs
}
