% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_data.R
\name{get_ab_alleles}
\alias{get_ab_alleles}
\title{Get A&B allele numbers from a list of CN data.}
\usage{
get_ab_alleles(mutation_data, cna_segments, sample_ids = NULL)
}
\arguments{
\item{mutation_data}{A CollapsedVCF object, GRanges object or a mutation id.}

\item{cna_segments}{A (named) list of CN segment data that contain A and B allele data.}

\item{sample_ids}{Indices of samples from 'cna_segments' to include in the output.}
}
\value{
A matrix containing annotations for each mutation object.
}
\description{
Get A&B allele numbers from a list of CN data.
}
