% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_data.R
\name{get_cnas}
\alias{get_cnas}
\title{Get CN annotations from a list of cn data.}
\usage{
get_cnas(mutation_data, cna_segments, sample_ids = NULL)
}
\arguments{
\item{mutation_data}{A CollapsedVCF object, GRanges object or a mutation id.}

\item{cna_segments}{A (named) list of CN segment data.}

\item{sample_ids}{Indices of samples from 'cna_segments' to include in the output.}
}
\value{
A matrix containing annotations for each mutation object.
}
\description{
Get CN annotations from a list of cn data.
}
