% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{get_mutation_df}
\alias{get_mutation_df}
\title{Get a data frame of mutation from various input formats}
\usage{
get_mutation_df(d)
}
\arguments{
\item{d}{A CollapsedVCF object or a mutation string.}
}
\value{
A data.frame containing mutation informations (chromosome, start, end and ref. & alt. alleles).
}
\description{
Get a data frame of mutation from various input formats
}
\examples{
get_mutation_df("chr1:757771_A/G")          # A single-nucleotide variant (SNV).
get_mutation_df("1:757771_A/G")             # Also a SNV.
get_mutation_df("chr1:757771_AT/AG")        # A multinucleotide variant.
get_mutation_df("chr1:757771-757772_AT/AG") # Also a MNV.
get_mutation_df("chr1:757771_ATG/A")        # A InDel.
get_mutation_df(c("chr1:757770_A/G","chr1:757771_ATG/A")) # Two mutations
}
