% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{get_mutation_type}
\alias{get_mutation_type}
\title{Get the mutation type of mutations}
\usage{
get_mutation_type(d)
}
\arguments{
\item{d}{A CollapsedVCF object or a mutation string.}
}
\value{
A named factor vector containing the mutation type (SNV, MNV or InDel).
}
\description{
Get the mutation type of mutations
}
\examples{
get_mutation_type("chr1:757771_A/G")          # A single-nucleotide variant (SNV).
get_mutation_type("1:757771_A/G")             # Also a SNV.
get_mutation_type("chr1:757771_AT/AG")        # A multinucleotide variant.
get_mutation_type("chr1:757771-757772_AT/AG") # Also a MNV.
get_mutation_type("chr1:757771_ATG/A")        # A InDel.
get_mutation_type(c("chr1:757770_A/G","chr1:757771_ATG/A")) # Two mutations
}
