% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{get_vaf}
\alias{get_vaf}
\title{Internal function to extract VAF information from any 'CollapsedVCF' object.}
\usage{
get_vaf(d)
}
\arguments{
\item{d}{Object of class 'CollapsedVCF'}
}
\value{
A numeric vector containing the VAF data.
}
\description{
Internal function to extract VAF information from any 'CollapsedVCF' object.
}
\keyword{internal}
