% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{guess_reference_genome}
\alias{guess_reference_genome}
\title{A internal function to guess the reference genome of a 'CollapsedVCF' object.}
\usage{
guess_reference_genome(d, verbose = TRUE)
}
\arguments{
\item{d}{A 'CollapsedVCF' object.}

\item{verbose}{Verbosity flag (Default: TRUE)}
}
\value{
Name of the matching reference genome
}
\description{
A internal function to guess the reference genome of a 'CollapsedVCF' object.
}
