% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{indentify_vcf_type}
\alias{indentify_vcf_type}
\title{Function to identify which tool created a given vcf file}
\usage{
indentify_vcf_type(f)
}
\arguments{
\item{f}{Path to a existing VCF file.}
}
\value{
A string indentifying the tool that (probably) created the file.
}
\description{
Function to identify which tool created a given vcf file
}
