% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{insert_cnas}
\alias{insert_cnas}
\title{Add an additional CNA and AB entries to a 'CollapsedVCF' object.}
\usage{
insert_cnas(mutation_data, cna_segments)
}
\arguments{
\item{mutation_data}{A 'CollapsedVCF' object}

\item{cna_segments}{A list of CNA data that can be parsed by the 'get_cnas' function}
}
\value{
A 'CollapsedVCF' object with the CNA and AB added to the genotype information
}
\description{
Add an additional CNA and AB entries to a 'CollapsedVCF' object.
}
