% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{insert_vaf}
\alias{insert_vaf}
\title{Adds an additional VAF entry to a 'CollapsedVCF' object.}
\usage{
insert_vaf(d)
}
\arguments{
\item{d}{A 'CollapsedVCF' object}
}
\value{
A 'CollapsedVCF' object with the VAF added to the genotype information
}
\description{
Adds an additional VAF entry to a 'CollapsedVCF' object.
}
