% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_between_references.R
\name{lift_data_wrapper}
\alias{lift_data_wrapper}
\title{Internal function used to lift mutation data between genome versions.}
\usage{
lift_data_wrapper(
  d,
  chain,
  geno,
  check_new_alleles = TRUE,
  multi_summary_function = NULL
)
}
\arguments{
\item{d}{Input data to work on. Options are i) a list of mutation vectors, ii) a GRanges object.}

\item{chain}{A chain file for the lifting.}

\item{geno}{A reference genome file of class 'BSgenome'.}

\item{check_new_alleles}{Logical flag indicating if match of reference alleles should be checked.}

\item{multi_summary_function}{Function used to summarise multiple segments in the new genome.}
}
\value{
A object with a structure similar to d with updated position data.
}
\description{
Internal function used to lift mutation data between genome versions.
}
