% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{load_mutect2_vcf}
\alias{load_mutect2_vcf}
\title{Internal wrapper function to load a Mutect2 vcf into a format usable for common purposes}
\usage{
load_mutect2_vcf(f)
}
\arguments{
\item{f}{A Platypus VCF file}
}
\value{
A object of class 'CollapsedVCF'
}
\description{
Internal wrapper function to load a Mutect2 vcf into a format usable for common purposes
}
\keyword{internal}
