% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_sequenza_purity_files}
\alias{load_sequenza_purity_files}
\title{Convience function to load Sequenza fit files.}
\usage{
load_sequenza_purity_files(files, verbose = FALSE)
}
\arguments{
\item{files}{Vector of files to load.}

\item{verbose}{Verbosity flag (Default: FALSE).}
}
\value{
Returns a data.frame containing per sample summary stats (purity, ploidy and average ploidy).
}
\description{
Convience function to load Sequenza fit files.
}
