% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_sequenza_segment_files}
\alias{load_sequenza_segment_files}
\title{Convience function to load Sequenza segment files.}
\usage{
load_sequenza_segment_files(files, verbose = FALSE)
}
\arguments{
\item{files}{Vector of files to load.}

\item{verbose}{Verbosity flag (Default: FALSE).}
}
\value{
Returns a list containing per sample data.
}
\description{
Convience function to load Sequenza segment files.
}
