% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{plot_cn_altered_per_patient}
\alias{plot_cn_altered_per_patient}
\title{Plot per-case copy-number alteration summary stats}
\usage{
plot_cn_altered_per_patient(data)
}
\arguments{
\item{data}{A list of summary stats returned by the 'calc_cn_summary_stats' function.}
}
\value{
A plot showing the fraction of samples showing copy-number alterations in each case.
}
\description{
Plot per-case copy-number alteration summary stats
}
