% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{plot_cn_freq}
\alias{plot_cn_freq}
\title{Plot copy-number alteration summary stats across cases}
\usage{
plot_cn_freq(
  data,
  cutoff_clonal = 1,
  cutoff_subclonal = 0,
  bands = NULL,
  gene_labels = NULL,
  ...
)
}
\arguments{
\item{data}{A list of summary stats returned by the 'calc_cn_summary_stats' function.}

\item{cutoff_clonal}{The fraction of samples that have to show a copy-number alteration in a case to consider it to be 'clonal' (i.e., present in all samples).}

\item{cutoff_subclonal}{The fraction of samples that have to show a copy-number alteration in a case to consider it to be 'subclonal' (i.e., present in some samples).}

\item{bands}{Optional information on chromosome banding. The format should be like those in the 'D3GB' package (e.g., D3GB::GRCh38.bands'), that is a data.frame containing 'chr', 'start', 'end' 'name' and 'score' columns. Rows with a 'score' of 'acen' will be used to calculate actual centromer positions.}

\item{gene_labels}{Optional gene names to label. Positions will be retrieved with the 'get_gene_pos' position.}

\item{...}{Optional arguments that will be passed to the 'get_gene_pos' function (see '?'get_gene_pos' for details).}
}
\value{
A plot showing the frequency of clonal and subclonal copy-number alterations across cases.
}
\description{
Plot copy-number alteration summary stats across cases
}
