% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{plot_cn_states_across_samples}
\alias{plot_cn_states_across_samples}
\title{Plot CN states of somatic sides across samples.}
\usage{
plot_cn_states_across_samples(data, min_freq = 0.01)
}
\arguments{
\item{data}{Object of class 'CollapsedVCF'}

\item{min_freq}{Minimum relative size to include a group of CN states. (Default: 0.01)}
}
\value{
A ggplot object containing mutation histograms
}
\description{
Plot CN states of somatic sides across samples.
}
