% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{plot_driver_vaf_heatmap}
\alias{plot_driver_vaf_heatmap}
\alias{save_driver_vaf_heatmap}
\title{Plot a heatmap showing VAF of specific mutations across samples.}
\usage{
plot_driver_vaf_heatmap(d, max_label_width = 50)

save_driver_vaf_heatmap(d, out_file, genes = NULL, ...)
}
\arguments{
\item{d}{Data frame containing the data to plot.}

\item{max_label_width}{A integer defining the maxmimum length of labels. Labels longer than this are trimmed.}

\item{out_file}{Name of the image file to create.}

\item{genes}{A character vector of genes to subset data to. (Default: NULL)}

\item{...}{Arguments passed to plot_driver_vaf_heatmap. See ?plot_driver_vaf_heatmap.}
}
\value{



}
\description{
Plot a heatmap showing VAF of specific mutations across samples.

Wrapper around plot_driver_vaf_heatmap function to save the image with resonable dimensions
}
