% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{plot_gene_mut_heatmap}
\alias{plot_gene_mut_heatmap}
\title{Create a plot showing if a gene is mutated in a given sample}
\usage{
plot_gene_mut_heatmap(
  d,
  crit_value = 0.1,
  genes = NULL,
  use = "CCF",
  order_alphabetic = FALSE,
  gene_to_gr = NULL,
  keep_all_genes = FALSE,
  xlab = ""
)
}
\arguments{
\item{d}{Data frame containing the data to plot.}

\item{crit_value}{Critical value of column 'use' to consider a sample to be mutated (i.e., d[,use] > crit_value)}

\item{genes}{# Genes to include in the plot.}

\item{use}{Column of the data frame containing information if a gene is mutated.}

\item{order_alphabetic}{Flag indicating if rows and columns should be ordered alphabetically.}

\item{gene_to_gr}{# Named character vector or labeller function used to convert gene ids to a group label.}

\item{keep_all_genes}{# Flag indicating if genes unmutaed in all samples should be included in the plot.}

\item{xlab}{# Label for the x-axis (e.g., the genes)}
}
\value{

}
\description{
Create a plot showing if a gene is mutated in a given sample
}
