% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{plot_inter_mutation_distance}
\alias{plot_inter_mutation_distance}
\title{Function to plot intermutation distances}
\usage{
plot_inter_mutation_distance(d, title = "", geno = NULL)
}
\arguments{
\item{d}{A mutation data frame. Can be obtained using 'vcf_to_data_frame' from a 'CollapsedVCF' object.}

\item{title}{Optional title to add to plot. . (Default: "")}

\item{geno}{A 'BSgenome' object that can be used to set limits and order of chromosomes. (Default: NULL)}
}
\value{
A ggplot object.
}
\description{
Function to plot intermutation distances
}
