% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{plot_pga_stats}
\alias{plot_pga_stats}
\title{Plot PGA stats}
\usage{
plot_pga_stats(pga_stats, layers = list())
}
\arguments{
\item{pga_stats}{data.frame containing data returned from `calc_pga_stats` (see `?calc_pga_stats` for details).}

\item{layers}{list of additional ggplot layers added to each subplot.}
}
\value{
A plot showing PGA stats.
}
\description{
Plot PGA stats
}
