% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvs.R
\name{plot_signature_data}
\alias{plot_signature_data}
\title{Plot signature spectrum of mutations}
\usage{
plot_signature_data(x)
}
\arguments{
\item{x}{Output of `annotate_mutation_types` or any input this function accepts.}

\item{...}{Additional arguments that would be passed to `annotate_mutation_types`.}
}
\value{
A plot of the signature spectrum.
}
\description{
Plot signature spectrum of mutations
}
