% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{plot_vaf_distribution}
\alias{plot_vaf_distribution}
\title{Plot a histogram of mutations contained in a VCF across CN states.}
\usage{
plot_vaf_distribution(d, what = "VAF", purity = NULL, ploidy = NULL)
}
\arguments{
\item{d}{Object of class 'CollapsedVCF'}

\item{what}{String containing 'VAF' or 'CCF' that defines what to plot. (Default: 'VAF')}

\item{purity}{A named numeric vector of sample purities. Used to calculate expected cluster positions and to create labels. (Default: NULL)}

\item{ploidy}{A named numeric vector of sample ploidies. Used to calculate expected cluster positions and to create labels. (Default: NULL)}
}
\value{
A ggplot object containing mutation histograms
}
\description{
Plot a histogram of mutations contained in a VCF across CN states.
}
