% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pretty_violin}
\alias{pretty_violin}
\title{Title}
\usage{
pretty_violin(color = "gray85", fill = "gray10", ...)
}
\arguments{
\item{color}{}

\item{fill}{}

\item{...}{}
}
\value{

}
\description{
Title
}
