% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print_file_info}
\alias{print_file_info}
\title{A internal function to print basic file statistics}
\usage{
print_file_info(f)
}
\arguments{
\item{f}{A file.}
}
\value{
NILL
}
\description{
A internal function to print basic file statistics
}
