% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{print_vcf_content}
\alias{print_vcf_content}
\title{A internal function that summarised the content of a 'CollapsedVCF' object.}
\usage{
print_vcf_content(d)
}
\arguments{
\item{d}{A 'CollapsedVCF' object.}
}
\value{
NILL
}
\description{
A internal function that summarised the content of a 'CollapsedVCF' object.
}
