% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{split_multiallelic}
\alias{split_multiallelic}
\title{Function to split multiallelic sites into multiple variants}
\usage{
split_multiallelic(d)
}
\arguments{
\item{d}{Object of class 'CollapsedVCF'}
}
\value{
Object of class 'CollapsedVCF' with multiallelic sites split
}
\description{
Function to split multiallelic sites into multiple variants
}
