% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_files.R
\name{vcf_to_data_frame}
\alias{vcf_to_data_frame}
\title{Convert a VCF to a data.frame containing key statistics}
\usage{
vcf_to_data_frame(
  d,
  relevant_consequences = c("high", "moderate"),
  annotate_all_variants = FALSE,
  annotate_all_genes = TRUE,
  include_annot = FALSE,
  trim_canocial = TRUE,
  genes = NULL
)
}
\arguments{
\item{d}{Object of class 'CollapsedVCF'}

\item{relevant_consequences}{A character vector defining the relevant consequences to return in the annotations. The possible options are 'high', 'moderate', 'low', 'modifier', 'all' or a combination of these. See https://www.ensembl.org/info/genome/variation/prediction/predicted_data.html for details. (Default: 'high' and 'moderate')}

\item{annotate_all_variants}{A logical flag indicating if all variants should be included in the annotations or if only the one of the canonical transcript should be keeped. (Default: FALSE)}

\item{annotate_all_genes}{A logical flag indicating if all genes should be included.  (Default: TRUE)}

\item{include_annot}{A logical flag indicating if the annotation of the variant should be included. (Default: FALSE)}

\item{trim_canocial}{A logical flag indicating if the transcript ID of canonical transcripts should be removed. (Default: TRUE)}

\item{genes}{A character vector of gene names to which the mutation data should be subset. If NULL all genes are included. (Default: NULL)}
}
\value{
A data.frame containing selected information.
}
\description{
Convert a VCF to a data.frame containing key statistics
}
