% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{window_list_from_genome}
\alias{window_list_from_genome}
\title{Creates a data.frame of named tiles from a BSgenome object}
\usage{
window_list_from_genome(geno, chr_filter = NULL, width = 1e+06)
}
\arguments{
\item{geno}{A BSgenome object}

\item{chr_filter}{Optional vector of chromosomes to subset to.}

\item{width}{The width of the tiles (Default: 1e6)}
}
\value{
A data.frame containing the positions of the tiles.
}
\description{
Creates a data.frame of named tiles from a BSgenome object
}
