\name{yrdif}
\alias{yrdif}
\title{Calculate difference between 2 dates in years.}
\usage{
  yrdif(startdate, stopdate, type = "actual")
}
\arguments{
  \item{startdate}{The first date}

  \item{stopdate}{The second date}

  \item{type}{TTTTTTT}
}
\description{
  Calculated the difference in years between two dates.
  The function calulates the number of days that fall in
  365-day years and divides that by 365.  Then adds the
  number of days that fall in 366-day years and divides
  that by 366.
}

