% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_analysis.R
\name{get_gene_pos}
\alias{get_gene_pos}
\title{Function to look-up (average) gene positions.}
\usage{
get_gene_pos(
  x,
  txdb = TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene,
  egdb = org.Hs.eg.db::org.Hs.eg.db,
  return_intervals = FALSE
)
}
\arguments{
\item{x}{}

\item{txdb}{}

\item{egdb}{}

\item{return_intervals}{}
}
\value{

}
\description{
Function to look-up (average) gene positions.
}
