% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pretty_violin}
\alias{pretty_violin}
\title{GGplot layer to plot a violing plot}
\usage{
pretty_violin(color = "gray85", fill = "gray10", ...)
}
\arguments{
\item{color}{Color of violin plot (default: light gray).}

\item{fill}{Color of overlayed boxplot (default: dark gray).}

\item{...}{unused}
}
\value{
A list containg ggplot layers to create a violin plot with a overlayed boxplot and mean and median points.
}
\description{
GGplot layer to plot a violing plot
}
